app_python/requirements-dev.txt
-r requirements.txt

asyncio==3.4.3
certifi==2024.2.2
httpcore==1.0.3
httpx==0.26.0
six==1.16.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
pytest==8.0.1
pytest-asyncio==0.23.5
python-dateutil==2.8.2